22/04/2023 - Carlos - Añadida funcionalidad para dar permisos de lectura a los amigos
