+++ webapp/src/components/pod/FriendsPOD.ts
@@ -4,7 +4,19 @@ import {
     getThing,
     getTermAll,
     IriString,
-    SolidDataset, Thing, buildThing, setThing, saveSolidDatasetAt,
+    SolidDataset,
+    Thing,
+    buildThing,
+    setThing,
+    saveSolidDatasetAt,
+    getSolidDatasetWithAcl,
+    hasResourceAcl,
+    hasAccessibleAcl,
+    hasFallbackAcl,
+    createAclFromFallbackAcl,
+    getResourceAcl,
+    setAgentResourceAccess,
+    setAgentDefaultAccess, saveAclFor,
 } from '@inrupt/solid-client'
 import { foaf, vcard, owl, rdfs } from 'rdf-namespaces'
 import {Session,fetch} from "@inrupt/solid-client-authn-browser";
@@ -109,4 +121,48 @@ export async function addFriendToPod(provider:string,friendName:string,webId:str
     solidDataset = setThing(solidDataset, friends);
     await saveSolidDatasetAt(webId, solidDataset, { fetch: fetch })
     return false
+}
+
+export async function changePermissions(webId: string, friendWebId: string,session:Session) {
+    // Fetch the SolidDataset and its associated ACLs, if available:
+    const myDatasetWithAcl = await getSolidDatasetWithAcl(webId + "private/", {fetch: session.fetch});
+
+    // Obtain the SolidDataset's own ACL, if available,
+    // or initialise a new one, if possible:
+    let resourceAcl;
+    if (!hasResourceAcl(myDatasetWithAcl)) {
+        if (!hasAccessibleAcl(myDatasetWithAcl)) {
+            throw new Error(
+                "The current user does not have permission to change access rights to this Resource."
+            );
+        }
+        if (!hasFallbackAcl(myDatasetWithAcl)) {
+            throw new Error(
+                "The current user does not have permission to see who currently has access to this Resource."
+            );
+            // Alternatively, initialise a new empty ACL as follows,
+            // but be aware that if you do not give someone Control access,
+            // **nobody will ever be able to change Access permissions in the future**:
+            // resourceAcl = createAcl(myDatasetWithAcl);
+        }
+        resourceAcl = createAclFromFallbackAcl(myDatasetWithAcl);
+    } else {
+        resourceAcl = getResourceAcl(myDatasetWithAcl);
+    }
+
+    // Give someone Control access to the given Resource:
+    let updatedAcl = setAgentResourceAccess(
+        resourceAcl,
+        friendWebId,
+        {read: true, append: false, write: false, control: false}
+    );
+    updatedAcl = setAgentDefaultAccess(
+        updatedAcl,
+        friendWebId,
+        {read: true, append: false, write: false, control: false}
+    )
+
+
+    // Now save the ACL:
+    await saveAclFor(myDatasetWithAcl, updatedAcl, {fetch: session.fetch});
 }
